More plantool fixes
